Line slope
This script is intended to calculate slopes of line layers, representing road networks, based on a Digital Terrain Model (DTM) in the form of a one band raster (or grid), with altitude values. The size of the cells should not exceed 50 meters, and a raster of 10 to 25 meters cell size is recommended.
The script does not change the original line layer, producing a new memory layer, called "Intersection", the latter being a line layer with all the fields of the original, plus a new field called "Slope", with the values of the slope in percentage (%).
It is advised that the line layer, DTM, and project are in the same projected coordinate system.

   1 - Input parameters
      1.1 - Segment length multiplier
The sections of the original layer are split into smaller segments, with a length equal to (Segment Length Multiplier x raster cell size).
The user must indicate a factor between 4 and 10, the default value being 5. Values less than 4 can produce serious errors in slope calculation, and values close to 10 lower the slope value accuracy for each segment, but have the advantage of increasing the speed of calculation.
       1.2 - Line layer
The original line layer (road network layer).
      1.3 - DTM
Digital Terrain Model (DTM) or DEM. Must be a one band raster (grid) with elevation values.

Outputs
A segmented line layer with an added field named "Slope" with percent slope values for each segment.